pyproject.toml
[project]
name = "yamlgrep"
version = "0.1.0"
description = "Dumb little tool to 'grep' yaml files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "rich-argparse>=1.7.1",
]   

[project.scripts]
yamlgrep = "yamlgrep.main:main"

[tool.uv]
package = true